working sketch of file delivery pattern
